Change Level field from int to float64 to support decimal values
The Level field on Exercise (used for cardio and other exercise types)
now accepts decimal/float values (e.g. 5.5) instead of whole numbers only.

+++ internal/models/exercise.go
@@ -34,7 +34,7 @@ type Exercise struct {
 	Time         int          `json:"time,omitempty"`
 	Distance     float64      `json:"distance,omitempty"`
 	DistanceUnit string       `json:"distanceUnit,omitempty"`
-	Level        int          `json:"level,omitempty"`
+	Level        float64      `json:"level,omitempty"`
 	Reps         int          `json:"reps,omitempty"`
 	Sets         []WeightItem `json:"sets,omitempty"`
 }
